refactor: Restructured ClientApp
