+++ ClientApp/src/Shared/Services/TextService.ts
