+++ ClientApp/src/Shared/Stores/authStore.ts
