--- ClientApp/src/assets/main.css
@@ -1,1 +1,0 @@
-@import './base.css';
